fix: [CI-4730]: refactor cie code (#34826)
