+++ 320-ci-execution/src/main/java/io/harness/ci/buildstate/CodebaseUtils.java
@@ -8,24 +8,24 @@
 package io.harness.ci.buildstate;
 
 import static io.harness.beans.sweepingoutputs.CISweepingOutputNames.CODEBASE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_AWS_REGION;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BUILD_EVENT;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_AVATAR;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_EMAIL;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_BEFORE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_REF;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_SHA;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_NETRC_MACHINE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_NETRC_PORT;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_NETRC_USERNAME;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REMOTE_URL;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_SOURCE_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_TAG;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_TARGET_BRANCH;
-import static io.harness.ci.common.CIExecutionConstants.AWS_CODE_COMMIT_URL_REGEX;
-import static io.harness.ci.common.CIExecutionConstants.PATH_SEPARATOR;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_AWS_REGION;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BUILD_EVENT;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_AVATAR;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_EMAIL;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_BEFORE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_REF;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_SHA;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_NETRC_MACHINE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_NETRC_PORT;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_NETRC_USERNAME;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REMOTE_URL;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_SOURCE_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_TAG;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_TARGET_BRANCH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.AWS_CODE_COMMIT_URL_REGEX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PATH_SEPARATOR;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 import static io.harness.delegate.beans.connector.ConnectorType.AZURE_REPO;
