+++ 320-ci-execution/src/main/java/io/harness/ci/buildstate/K8BuildSetupUtils.java
@@ -16,31 +16,31 @@ import static io.harness.beans.sweepingoutputs.CISweepingOutputNames.CODE_BASE_C
 import static io.harness.beans.sweepingoutputs.ContainerPortDetails.PORT_DETAILS;
 import static io.harness.beans.sweepingoutputs.PodCleanupDetails.CLEANUP_DETAILS;
 import static io.harness.beans.sweepingoutputs.StageInfraDetails.STAGE_INFRA_DETAILS;
-import static io.harness.ci.common.CIExecutionConstants.ACCOUNT_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.BUILD_NUMBER_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_LOG_PREFIX_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_PIPELINE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_SERVICE_LOG_KEY_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_WORKSPACE;
-import static io.harness.ci.common.CIExecutionConstants.LABEL_REGEX;
-import static io.harness.ci.common.CIExecutionConstants.LOG_SERVICE_ENDPOINT_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.LOG_SERVICE_TOKEN_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.ORG_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.PIPELINE_EXECUTION_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.PIPELINE_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.POD_MAX_WAIT_UNTIL_READY_SECS;
-import static io.harness.ci.common.CIExecutionConstants.PROJECT_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.STAGE_ID_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.STAGE_NAME_ATTR;
-import static io.harness.ci.common.CIExecutionConstants.TI_SERVICE_ENDPOINT_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.TI_SERVICE_TOKEN_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ACCOUNT_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.BUILD_NUMBER_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_LOG_PREFIX_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_PIPELINE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_SERVICE_LOG_KEY_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_WORKSPACE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LABEL_REGEX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LOG_SERVICE_ENDPOINT_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LOG_SERVICE_TOKEN_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ORG_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PIPELINE_EXECUTION_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PIPELINE_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.POD_MAX_WAIT_UNTIL_READY_SECS;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PROJECT_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STAGE_ID_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STAGE_NAME_ATTR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.TI_SERVICE_ENDPOINT_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.TI_SERVICE_TOKEN_VARIABLE;
 import static io.harness.common.STOExecutionConstants.STO_SERVICE_ENDPOINT_VARIABLE;
 import static io.harness.common.STOExecutionConstants.STO_SERVICE_TOKEN_VARIABLE;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
