+++ 320-ci-execution/src/main/java/io/harness/ci/buildstate/ServiceContainerUtils.java
@@ -5,11 +5,14 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ci.buildstate.providers;
+package io.harness.ci.buildstate;
 
-import static io.harness.ci.common.CIExecutionConstants.PORT_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.UNIX_STEP_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.WIN_STEP_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ID_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.IMAGE_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PORT_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SERVICE_ARG_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.UNIX_STEP_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.WIN_STEP_COMMAND;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
@@ -19,7 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @OwnedBy(HarnessTeam.CI)
-public class StepContainerUtils {
+public class ServiceContainerUtils {
   public static List<String> getCommand(OSType os) {
     String cmd = UNIX_STEP_COMMAND;
     if (os == OSType.Windows) {
@@ -31,8 +34,14 @@ public class StepContainerUtils {
     return command;
   }
 
-  public static List<String> getArguments(Integer port) {
+  public static List<String> getArguments(String serviceID, String image, Integer port) {
     List<String> args = new ArrayList<>();
+    args.add(SERVICE_ARG_COMMAND);
+    args.add(ID_PREFIX);
+    args.add(serviceID);
+    args.add(IMAGE_PREFIX);
+    args.add(image);
+
     args.add(PORT_PREFIX);
     args.add(port.toString());
     return args;
