+++ 320-ci-execution/src/main/java/io/harness/ci/buildstate/StepContainerUtils.java
@@ -5,14 +5,11 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ci.buildstate.providers;
+package io.harness.ci.buildstate;
 
-import static io.harness.ci.common.CIExecutionConstants.ID_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.IMAGE_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.PORT_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.SERVICE_ARG_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.UNIX_STEP_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.WIN_STEP_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PORT_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.UNIX_STEP_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.WIN_STEP_COMMAND;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
@@ -22,7 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @OwnedBy(HarnessTeam.CI)
-public class ServiceContainerUtils {
+public class StepContainerUtils {
   public static List<String> getCommand(OSType os) {
     String cmd = UNIX_STEP_COMMAND;
     if (os == OSType.Windows) {
@@ -34,14 +31,8 @@ public class ServiceContainerUtils {
     return command;
   }
 
-  public static List<String> getArguments(String serviceID, String image, Integer port) {
+  public static List<String> getArguments(Integer port) {
     List<String> args = new ArrayList<>();
-    args.add(SERVICE_ARG_COMMAND);
-    args.add(ID_PREFIX);
-    args.add(serviceID);
-    args.add(IMAGE_PREFIX);
-    args.add(image);
-
     args.add(PORT_PREFIX);
     args.add(port.toString());
     return args;
