+++ 320-ci-execution/src/main/java/io/harness/ci/buildstate/providers/InternalContainerParamsProvider.java
@@ -7,27 +7,27 @@
 
 package io.harness.ci.buildstate.providers;
 
-import static io.harness.ci.common.CIExecutionConstants.DELEGATE_SERVICE_ENDPOINT_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.DELEGATE_SERVICE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.DELEGATE_SERVICE_ID_VARIABLE_VALUE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_LOG_PREFIX_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_PIPELINE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_WORKSPACE;
-import static io.harness.ci.common.CIExecutionConstants.LITE_ENGINE_CONTAINER_CPU;
-import static io.harness.ci.common.CIExecutionConstants.LITE_ENGINE_CONTAINER_MEM;
-import static io.harness.ci.common.CIExecutionConstants.LITE_ENGINE_CONTAINER_NAME;
-import static io.harness.ci.common.CIExecutionConstants.PWSH_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.SETUP_ADDON_CONTAINER_NAME;
-import static io.harness.ci.common.CIExecutionConstants.SH_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.UNIX_SETUP_ADDON_ARGS;
-import static io.harness.ci.common.CIExecutionConstants.WIN_SETUP_ADDON_ARGS;
+import static io.harness.ci.commonconstants.CIExecutionConstants.DELEGATE_SERVICE_ENDPOINT_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.DELEGATE_SERVICE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.DELEGATE_SERVICE_ID_VARIABLE_VALUE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_LOG_PREFIX_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_PIPELINE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_WORKSPACE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LITE_ENGINE_CONTAINER_CPU;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LITE_ENGINE_CONTAINER_MEM;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LITE_ENGINE_CONTAINER_NAME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PWSH_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SETUP_ADDON_CONTAINER_NAME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SH_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.UNIX_SETUP_ADDON_ARGS;
+import static io.harness.ci.commonconstants.CIExecutionConstants.WIN_SETUP_ADDON_ARGS;
 import static io.harness.data.encoding.EncodingUtils.encodeBase64;
 import static io.harness.delegate.beans.ci.pod.SecretParams.Type.TEXT;
