+++ 320-ci-execution/src/main/java/io/harness/ci/commonconstants/BuildEnvironmentConstants.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ci.common;
+package io.harness.ci.commonconstants;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
