+++ 320-ci-execution/src/main/java/io/harness/ci/commonconstants/CICommonPodConstants.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.ci.common;
+package io.harness.ci.commonconstants;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
@@ -14,8 +14,9 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 @OwnedBy(HarnessTeam.CI)
-public class CICommonEndpointConstants {
-  public static final String LOG_SERVICE_TOKEN_ENDPOINT = "token";
-  public static final String LOG_SERVICE_STREAM_ENDPOINT = "stream";
-  public static final String TI_SERVICE_TOKEN_ENDPOINT = "token";
+public class CICommonPodConstants {
+  public static final String POD_NAME_PREFIX = "harnessci";
+  public static final String CONTAINER_NAME = "build-setup";
+  public static final String REL_STDOUT_FILE_PATH = "/stdout";
+  public static final String REL_STDERR_FILE_PATH = "/stderr";
 }
