+++ 320-ci-execution/src/main/java/io/harness/ci/integrationstage/BuildEnvironmentUtils.java
@@ -7,35 +7,35 @@
 
 package io.harness.ci.integrationstage;
 
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BUILD_ACTION;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BUILD_EVENT;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BUILD_NUMBER;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AFTER;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_AVATAR;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_EMAIL;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_NAME;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_BEFORE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_LINK;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_MESSAGE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_REF;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_SHA;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_GIT_HTTP_URL;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_GIT_SSH_URL;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_LINK;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_NAME;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_NAMESPACE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_OWNER;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_PRIVATE;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REPO_SCM;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_SOURCE_BRANCH;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_TAG;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_TARGET_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BUILD_ACTION;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BUILD_EVENT;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BUILD_NUMBER;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AFTER;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_AVATAR;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_EMAIL;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_AUTHOR_NAME;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_BEFORE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_LINK;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_MESSAGE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_REF;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_SHA;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_GIT_HTTP_URL;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_GIT_SSH_URL;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_LINK;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_NAME;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_NAMESPACE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_OWNER;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_PRIVATE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REPO_SCM;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_SOURCE_BRANCH;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_TAG;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_TARGET_BRANCH;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
