+++ 320-ci-execution/src/main/java/io/harness/ci/integrationstage/K8InitializeServiceUtils.java
@@ -10,9 +10,9 @@ package io.harness.ci.integrationstage;
 import static io.harness.beans.serializer.RunTimeInputHandler.UNRESOLVED_PARAMETER;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveIntegerParameter;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveStringParameter;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_SERVICE_ARGS;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_SERVICE_ENTRYPOINT;
-import static io.harness.ci.common.CIExecutionConstants.SERVICE_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_SERVICE_ARGS;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_SERVICE_ENTRYPOINT;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SERVICE_PREFIX;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 
@@ -28,7 +28,7 @@ import io.harness.beans.quantity.unit.StorageQuantityUnit;
 import io.harness.beans.serializer.RunTimeInputHandler;
 import io.harness.beans.stages.IntegrationStageConfig;
 import io.harness.beans.yaml.extended.infrastrucutre.OSType;
-import io.harness.ci.buildstate.providers.ServiceContainerUtils;
+import io.harness.ci.buildstate.ServiceContainerUtils;
 import io.harness.ci.config.CIExecutionServiceConfig;
 import io.harness.ci.execution.CIExecutionConfigService;
 import io.harness.ci.utils.PortFinder;
