+++ 320-ci-execution/src/main/java/io/harness/ci/integrationstage/K8InitializeStepInfoBuilder.java
@@ -13,22 +13,22 @@ import static io.harness.beans.serializer.RunTimeInputHandler.resolveMapParamete
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveOSType;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveStringParameter;
 import static io.harness.beans.serializer.RunTimeInputHandler.resolveStringParameterWithDefaultValue;
-import static io.harness.ci.common.CICommonPodConstants.POD_NAME_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.ADDON_VOLUME;
-import static io.harness.ci.common.CIExecutionConstants.ADDON_VOL_MOUNT_PATH;
-import static io.harness.ci.common.CIExecutionConstants.DEFAULT_CONTAINER_CPU_POV;
-import static io.harness.ci.common.CIExecutionConstants.DEFAULT_CONTAINER_MEM_POV;
-import static io.harness.ci.common.CIExecutionConstants.ETC_DIR;
-import static io.harness.ci.common.CIExecutionConstants.PORT_STARTING_RANGE;
-import static io.harness.ci.common.CIExecutionConstants.PVC_DEFAULT_STORAGE_CLASS;
-import static io.harness.ci.common.CIExecutionConstants.SHARED_VOLUME_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.STEP_MOUNT_PATH;
-import static io.harness.ci.common.CIExecutionConstants.STEP_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.STEP_REQUEST_MEMORY_MIB;
-import static io.harness.ci.common.CIExecutionConstants.STEP_REQUEST_MILLI_CPU;
-import static io.harness.ci.common.CIExecutionConstants.STEP_VOLUME;
-import static io.harness.ci.common.CIExecutionConstants.STEP_WORK_DIR;
-import static io.harness.ci.common.CIExecutionConstants.VOLUME_PREFIX;
+import static io.harness.ci.commonconstants.CICommonPodConstants.POD_NAME_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ADDON_VOLUME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ADDON_VOL_MOUNT_PATH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.DEFAULT_CONTAINER_CPU_POV;
+import static io.harness.ci.commonconstants.CIExecutionConstants.DEFAULT_CONTAINER_MEM_POV;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ETC_DIR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PORT_STARTING_RANGE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PVC_DEFAULT_STORAGE_CLASS;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SHARED_VOLUME_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_MOUNT_PATH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_REQUEST_MEMORY_MIB;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_REQUEST_MILLI_CPU;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_VOLUME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_WORK_DIR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.VOLUME_PREFIX;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
 
@@ -65,7 +65,7 @@ import io.harness.beans.yaml.extended.volumes.HostPathYaml;
 import io.harness.beans.yaml.extended.volumes.PersistentVolumeClaimYaml;
 import io.harness.ci.buildstate.ConnectorUtils;
 import io.harness.ci.buildstate.PluginSettingUtils;
-import io.harness.ci.buildstate.providers.StepContainerUtils;
+import io.harness.ci.buildstate.StepContainerUtils;
 import io.harness.ci.execution.CIExecutionConfigService;
 import io.harness.ci.ff.CIFeatureFlagService;
 import io.harness.ci.utils.CIStepInfoUtils;
