+++ 320-ci-execution/src/main/java/io/harness/ci/states/AbstractStepExecutable.java
@@ -11,8 +11,8 @@ import static io.harness.annotations.dev.HarnessTeam.CI;
 import static io.harness.beans.sweepingoutputs.CISweepingOutputNames.CODE_BASE_CONNECTOR_REF;
 import static io.harness.beans.sweepingoutputs.ContainerPortDetails.PORT_DETAILS;
 import static io.harness.beans.sweepingoutputs.StageInfraDetails.STAGE_INFRA_DETAILS;
-import static io.harness.ci.common.CIExecutionConstants.LITE_ENGINE_PORT;
-import static io.harness.ci.common.CIExecutionConstants.TMP_PATH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LITE_ENGINE_PORT;
+import static io.harness.ci.commonconstants.CIExecutionConstants.TMP_PATH;
 import static io.harness.ci.states.InitializeTaskStep.LE_STATUS_TASK_TYPE;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
