+++ 320-ci-execution/src/test/java/io/harness/ci/executionplan/CIExecutionPlanTestHelper.java
@@ -8,22 +8,22 @@
 package io.harness.ci.executionplan;
 
 import static io.harness.annotations.dev.HarnessTeam.CI;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_BUILD_NUMBER;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
-import static io.harness.ci.common.CIExecutionConstants.GIT_CLONE_DEPTH_ATTRIBUTE;
-import static io.harness.ci.common.CIExecutionConstants.GIT_CLONE_MANUAL_DEPTH;
-import static io.harness.ci.common.CIExecutionConstants.GIT_CLONE_STEP_ID;
-import static io.harness.ci.common.CIExecutionConstants.GIT_CLONE_STEP_NAME;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_WORKSPACE;
-import static io.harness.ci.common.CIExecutionConstants.ID_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.IMAGE_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.PLUGIN_ENV_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.PORT_PREFIX;
-import static io.harness.ci.common.CIExecutionConstants.PORT_STARTING_RANGE;
-import static io.harness.ci.common.CIExecutionConstants.SERVICE_ARG_COMMAND;
-import static io.harness.ci.common.CIExecutionConstants.STEP_REQUEST_MEMORY_MIB;
-import static io.harness.ci.common.CIExecutionConstants.STEP_REQUEST_MILLI_CPU;
-import static io.harness.ci.common.CIExecutionConstants.UNIX_STEP_COMMAND;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_BUILD_NUMBER;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_COMMIT_BRANCH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.GIT_CLONE_DEPTH_ATTRIBUTE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.GIT_CLONE_MANUAL_DEPTH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.GIT_CLONE_STEP_ID;
+import static io.harness.ci.commonconstants.CIExecutionConstants.GIT_CLONE_STEP_NAME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_WORKSPACE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ID_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.IMAGE_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PLUGIN_ENV_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PORT_PREFIX;
+import static io.harness.ci.commonconstants.CIExecutionConstants.PORT_STARTING_RANGE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SERVICE_ARG_COMMAND;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_REQUEST_MEMORY_MIB;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_REQUEST_MILLI_CPU;
+import static io.harness.ci.commonconstants.CIExecutionConstants.UNIX_STEP_COMMAND;
 import static io.harness.delegate.beans.ci.pod.CIContainerType.PLUGIN;
 import static io.harness.delegate.beans.ci.pod.CIContainerType.RUN;
 import static io.harness.delegate.beans.ci.pod.CIContainerType.SERVICE;
@@ -68,7 +68,7 @@ import io.harness.beans.yaml.extended.volumes.HostPathYaml;
 import io.harness.beans.yaml.extended.volumes.HostPathYaml.HostPathYamlSpec;
 import io.harness.beans.yaml.extended.volumes.PersistentVolumeClaimYaml;
 import io.harness.beans.yaml.extended.volumes.PersistentVolumeClaimYaml.PersistentVolumeClaimYamlSpec;
-import io.harness.ci.common.CIExecutionConstants;
+import io.harness.ci.commonconstants.CIExecutionConstants;
 import io.harness.connector.ConnectorDTO;
 import io.harness.connector.ConnectorInfoDTO;
 import io.harness.delegate.beans.ci.pod.CIK8ContainerParams;
