+++ 320-ci-execution/src/test/java/io/harness/ci/stateutils/buildstate/K8BuildSetupUtilsTest.java
@@ -7,24 +7,24 @@
 
 package io.harness.ci.stateutils.buildstate;
 
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_AWS_REGION;
-import static io.harness.ci.common.BuildEnvironmentConstants.DRONE_REMOTE_URL;
-import static io.harness.ci.common.CIExecutionConstants.ACCESS_KEY_MINIO_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.LOG_SERVICE_ENDPOINT_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.LOG_SERVICE_TOKEN_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.SECRET_KEY_MINIO_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.STEP_MOUNT_PATH;
-import static io.harness.ci.common.CIExecutionConstants.STEP_VOLUME;
-import static io.harness.ci.common.CIExecutionConstants.STEP_WORK_DIR;
-import static io.harness.ci.common.CIExecutionConstants.TI_SERVICE_ENDPOINT_VARIABLE;
-import static io.harness.ci.common.CIExecutionConstants.TI_SERVICE_TOKEN_VARIABLE;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_AWS_REGION;
+import static io.harness.ci.commonconstants.BuildEnvironmentConstants.DRONE_REMOTE_URL;
+import static io.harness.ci.commonconstants.CIExecutionConstants.ACCESS_KEY_MINIO_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ACCOUNT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_BUILD_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_CI_INDIRECT_LOG_UPLOAD_FF;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_EXECUTION_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_ORG_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_PROJECT_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.HARNESS_STAGE_ID_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LOG_SERVICE_ENDPOINT_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.LOG_SERVICE_TOKEN_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.SECRET_KEY_MINIO_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_MOUNT_PATH;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_VOLUME;
+import static io.harness.ci.commonconstants.CIExecutionConstants.STEP_WORK_DIR;
+import static io.harness.ci.commonconstants.CIExecutionConstants.TI_SERVICE_ENDPOINT_VARIABLE;
+import static io.harness.ci.commonconstants.CIExecutionConstants.TI_SERVICE_TOKEN_VARIABLE;
 import static io.harness.common.STOExecutionConstants.STO_SERVICE_ENDPOINT_VARIABLE;
 import static io.harness.common.STOExecutionConstants.STO_SERVICE_TOKEN_VARIABLE;
 import static io.harness.data.structure.UUIDGenerator.generateUuid;
